+++ java-client/src/test/java/org/openapitools/client/api/NamespaceBranchManagementApiTest.java
@@ -162,8 +162,9 @@ public class NamespaceBranchManagementApiTest {
         String clusterName = null;
         String namespaceName = null;
         String branchName = null;
-        String operator = null;
         OpenGrayReleaseRuleDTO openGrayReleaseRuleDTO = null;
+        String operator = null;
+        api.updateBranchRules(appId, env, clusterName, namespaceName, branchName, openGrayReleaseRuleDTO, operator);
         api.updateBranchRules(appId, env, clusterName, namespaceName, branchName, operator, openGrayReleaseRuleDTO);
         // TODO: test validations
     }