+++ python/apollo_openapi/paths/openapi_v1_envs_env_apps_app_id_clusters_cluster_name_namespaces_namespace_name_branches_branch_name/delete.py
@@ -31,12 +31,12 @@ OperatorSchema = schemas.StrSchema
 RequestRequiredQueryParams = typing_extensions.TypedDict(
     'RequestRequiredQueryParams',
     {
-        'operator': typing.Union[OperatorSchema, str, ],
     }
 )
 RequestOptionalQueryParams = typing_extensions.TypedDict(
     'RequestOptionalQueryParams',
     {
+        'operator': typing.Union[OperatorSchema, str, ],
     },
     total=False
 )
@@ -50,7 +50,6 @@ request_query_operator = api_client.QueryParameter(
     name="operator",
     style=api_client.ParameterStyle.FORM,
     schema=OperatorSchema,
-    required=True,
     explode=True,
 )
 # Path params